+++ spec/lib/hitnmiss/repository_spec.rb
@@ -40,17 +40,28 @@ describe Hitnmiss::Repository do
     end
   end
 
-  describe ".perform" do
+  describe ".fetch_cacheable_entity" do
     it "raises error indicating not implemented" do
       repo_klass = Class.new do
         include Hitnmiss::Repository
       end
 
-      expect { repo_klass.perform }.to raise_error(Hitnmiss::Errors::NotImplemented)
+      expect { repo_klass.fetch_cacheable_entity }.to raise_error(Hitnmiss::Errors::NotImplemented)
     end
   end
 
-  describe ".prime_cache" do
+  describe ".fetch_cacheable_entities" do
+    it "raises error indicating not implemented" do
+      repo_klass = Class.new do
+        include Hitnmiss::Repository
+      end
+
+      keyspace = double('keyspace')
+      expect { repo_klass.fetch_cacheable_entities(keyspace) }.to raise_error(Hitnmiss::Errors::NotImplemented)
+    end
+  end
+
+  describe ".prime_entity_cache" do
     it "obtains the cacheable entity" do
       repo_klass = Class.new do
         include Hitnmiss::Repository
@@ -59,9 +70,9 @@ describe Hitnmiss::Repository do
 
       args = double('arguments')
       Hitnmiss.register_driver(:my_driver, double.as_null_object)
-      expect(repo_klass).to receive(:perform).with(args).and_return(double.as_null_object)
+      expect(repo_klass).to receive(:fetch_cacheable_entity).with(args).and_return(double.as_null_object)
 
-      repo_klass.prime_cache(args)
+      repo_klass.prime_entity_cache(args)
     end
 
     context "when cacheable entity has an expiration" do
@@ -74,7 +85,7 @@ describe Hitnmiss::Repository do
           include Hitnmiss::Repository
           self.driver :my_driver
 
-          def self.perform(*args)
+          def self.fetch_cacheable_entity(*args)
             Hitnmiss::Entity.new('myval', 22223)
           end
         end
@@ -84,7 +95,7 @@ describe Hitnmiss::Repository do
 
         expect(driver).to receive(:set).with(key, 'myval', 22223)
 
-        repo_klass.prime_cache(args)
+        repo_klass.prime_entity_cache(args)
       end
     end
 
@@ -99,7 +110,7 @@ describe Hitnmiss::Repository do
           include Hitnmiss::Repository
           self.driver :my_driver
 
-          def self.perform(*args)
+          def self.fetch_cacheable_entity(*args)
             Hitnmiss::Entity.new('myval')
           end
         end
@@ -111,7 +122,7 @@ describe Hitnmiss::Repository do
 
         expect(driver).to receive(:set).with(key, 'myval', default_expiration)
 
-        repo_klass.prime_cache(args)
+        repo_klass.prime_entity_cache(args)
       end
     end
 
@@ -123,9 +134,39 @@ describe Hitnmiss::Repository do
       entity = double(value: 'foovalue', expiration: 212).as_null_object
 
       args = double('arguments')
-      allow(repo_klass).to receive(:perform).with(args).and_return(entity)
+      allow(repo_klass).to receive(:fetch_cacheable_entity).with(args).and_return(entity)
 
-      expect(repo_klass.prime_cache(args)).to eq('foovalue')
+      expect(repo_klass.prime_entity_cache(args)).to eq('foovalue')
+    end
+  end
+
+  describe ".prime_cache" do
+    it 'caches cacheable entities' do
+      key1 = double('key 1')
+      key2 = double('key 2')
+      driver = double('cache driver')
+      args = double('arguments')
+
+      repo_klass = Class.new do
+        include Hitnmiss::Repository
+        self.driver :my_driver
+
+        def self.fetch_cacheable_entities(keyspace)
+          [
+            { args: ['key1'], entity: Hitnmiss::Entity.new('myval', 22223) },
+            { args: ['key2'], entity: Hitnmiss::Entity.new('myval2', 43564) }
+          ]
+        end
+      end
+      Hitnmiss.register_driver(:my_driver, driver)
+
+      allow(repo_klass).to receive(:generate_key).with('key1').and_return(key1)
+      allow(repo_klass).to receive(:generate_key).with('key2').and_return(key2)
+
+      expect(driver).to receive(:set).with(key1, 'myval', 22223)
+      expect(driver).to receive(:set).with(key2, 'myval2', 43564)
+
+      repo_klass.prime_cache
     end
   end
 
@@ -153,7 +194,7 @@ describe Hitnmiss::Repository do
       Hitnmiss.register_driver(:my_driver, driver)
       key = double('key')
       allow(repo_klass).to receive(:generate_key).and_return(key)
-      allow(repo_klass).to receive(:prime_cache)
+      allow(repo_klass).to receive(:prime_entity_cache)
 
       expect(driver).to receive(:get).with(key)
 
@@ -212,7 +253,7 @@ describe Hitnmiss::Repository do
         allow(repo_klass).to receive(:generate_key).and_return(key)
         Hitnmiss.register_driver(:my_driver, driver)
         allow(driver).to receive(:get).with(key).and_return(nil)
-        expect(repo_klass).to receive(:prime_cache).with('aoeuaoeuao')
+        expect(repo_klass).to receive(:prime_entity_cache).with('aoeuaoeuao')
 
         repo_klass.fetch('aoeuaoeuao')
       end
@@ -228,10 +269,57 @@ describe Hitnmiss::Repository do
         allow(repo_klass).to receive(:generate_key).and_return(key)
         Hitnmiss.register_driver(:my_driver, driver)
         allow(driver).to receive(:get).with(key).and_return(nil)
-        allow(repo_klass).to receive(:prime_cache).and_return('porkpork')
+        allow(repo_klass).to receive(:prime_entity_cache).and_return('porkpork')
 
         expect(repo_klass.fetch('aoeuaoeuao')).to eq('porkpork')
       end
+    end
+  end
+
+  describe '.delete' do
+    it 'generates the cache key' do
+      repo_klass = Class.new do
+        include Hitnmiss::Repository
+        self.driver :my_driver
+      end
+
+      driver = double('cache driver', del: nil)
+      expect(repo_klass).to receive(:generate_key).with('auaeuaoeua')
+      Hitnmiss.register_driver(:my_driver, driver)
+
+      repo_klass.delete('auaeuaoeua')
+    end
+
+    it 'deletes the cached value' do
+      repo_klass = Class.new do
+        include Hitnmiss::Repository
+        driver :my_driver
+      end
+
+      driver = double('cache driver')
+      Hitnmiss.register_driver(:my_driver, driver)
+      key = double('key')
+      allow(repo_klass).to receive(:generate_key).and_return(key)
+
+      expect(driver).to receive(:del).with(key)
+
+      repo_klass.delete('aoeuaoeuao')
+    end
+  end
+
+  describe '.all' do
+    it 'returns all values for the keyspace' do
+      repo_klass = Class.new do
+        include Hitnmiss::Repository
+        self.driver :my_driver
+      end
+
+      entities = double('collection of entities')
+
+      driver = double('cache driver', all: entities)
+      Hitnmiss.register_driver(:my_driver, driver)
+
+      expect(repo_klass.all).to eq entities
     end
   end
 end